requirements.txt
pandas
streamlit
mysql.connector
